pyproject.toml
[tool.poetry]
name = "group03052024"
version = "0.1.0"
description = ""
authors = ["Василь Картичак <w.i.k.mailua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
